+++ src/include/Scene.h
@@ -40,8 +40,8 @@ class Scene
     void simulate(const size_t frames);
     void registerAction(int inputKey, const std::string& actionName);
 
-    size_t width() const;
-    size_t height() const;
+    int width() const;
+    int height() const;
     size_t currentFrame() const;
 
     bool hasEnded() const;
